feat: Zone Type 컴포넌트 분리
